Feat(Movies): Add list virtualization to movies list
* Refactors CustomTabs to take up parent height.
  This is so that AutoSizer will consume
  the available tab panel space.
* Adds lodash.chunk dependency
* Adds react-virtualized-auto-sizer dependency
* Adds virtua dependency for list virtualization

List virtualization is implemented with
virtua library, which supports dynamically
sized items out of the box.
Uses react-virtualized-auto-sizer to
create the list window.

Movies are manually chunked since
it's not a static grid, and dynamic grids
in the virtualization libraries
are suboptimal for this use case.
So we use a standard vertical list
for virtualization and chunk movies
into rows manually.